autoblitz-backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
boto3==1.29.7
redis==5.0.1
celery==5.3.4
python-jose[cryptography]==3.3.0
httpx==0.25.2
websockets==12.0
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic-settings==2.1.0
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
aiosqlite==0.19.0
aiosqlite==0.19.0
ccxt==4.3.74
websockets==12.0
aiohttp==3.9.5
cryptography==41.0.7
python-dotenv==1.0.0
pydantic-settings>=2.0.0
